add logic in CaculdateOutley, updata bankcontroler, update model,dto

+++ DailySpendServer/Model/BankAccount.cs
@@ -8,7 +8,7 @@
         public string Name { get; set; } = string.Empty;
         public string WebHookUrl { get; set; } = string.Empty;
         public string CashbackType { get; set; } = string.Empty;
-        public int Balance { get; set; }
+        public long Balance { get; set; }
         public string cardId { get; set; } = string.Empty;
 
         public string MaskedPan { get; set; } = string.Empty;
